new_position_bot/requirements.txt
# requirements.txt
functions-framework==3.8.2
requests==2.31.0
cryptography==41.0.7
google-cloud-secret-manager==2.16.0
pytest==7.4.3
python-dotenv==1.0.0
